Install libsrtp, libnice, and libav from rpm, not source
libnice for centos is 0.1.3 and licode needs 0.1.4 with a patch, so
we're building our own rpm and serving it from our internal yum server.

We build an rpm of libsrtp with the diffs between the licode version and
the upstream version so we no longer need to import the libsrtp source
from the libsrtp CVS repo.  Checking a cvs working copy into git is
pretty tacky so it's good that we don't need that anymore.

Both libnice and libsrtp had spec files in Centos, but libav didn't
since Red Hat distros use ffmpeg.  I found a spec file on github and
modified it to work with Centos.

All three spec file projects are in our github account: rpms-lib*.