Have erizoController serve erizo.js
This allows client apps to use erizo without having to install erizo.js
in their apps.  erizo.js is served from the same address that clients
use to rendez-vous with erizo.

For example, if erizo were running on 192.168.33.99 port 8080 then the
url for erizo.js would be http://192.168.33.99:8080/erizo.js and the url
for the debug version would be http://192.168.33.99:8080/erizo-debug.js.